Change filename from Boilerplate to CaverExample
