+++ abi/dynamic_abi_loader/src/main/java/CaverExample.java
@@ -17,11 +17,11 @@ import java.nio.file.Paths;
 import java.util.List;
 
 /**
- * Boilerplate code about "How to execute Smart Contract."
- * Related reference - Korean: https://ko.docs.klaytn.com/bapp/sdk/caver-java/getting-started#smart-contract
- * Related reference - English: https://docs.klaytn.com/bapp/sdk/caver-java/getting-started#smart-contract
+ * Example code about "Dynamic ABI Loader"
+ * Related article - Korean: https://medium.com/klaytn/caver-java-dynamic-abi-loader-8ccef10e7e68
+ * Related article - English: https://medium.com/klaytn/caver-java-dynamic-abi-loader-5b116c9db673
  */
-public class Boilerplate {
+public class CaverExample {
     // You can directly input values for the variables below, or you can enter values in the caver-java-examples/.env file.
     private static String nodeApiUrl = ""; // e.g. "https://node-api.klaytnapi.com/v1/klaytn";
     private static String accessKeyId = ""; // e.g. "KASK1LVNO498YT6KJQFUPY8S";
@@ -29,7 +29,6 @@ public class Boilerplate {
     private static String chainId = ""; // e.g. "1001" or "8217";
     private static String deployerAddress = ""; // e.g. "0xf5d3322418f2f2257d9255ad4a7ff1d50312d438"
     private static String deployerPrivateKey = ""; // e.g. "0x42f6375b608c2572fadb2ed9fd78c5c456ca3aa860c43192ad910c3269727fc7"
-
 
     public static void main(String[] args) {
         try {
@@ -72,7 +71,7 @@ public class Boilerplate {
         httpService.addHeader("x-chain-id", chainId);
         Caver caver = new Caver(httpService);
 
-        // abi is extracted by compiling caver-java-examples/resources/KVstore.sol using solc(solidity compiler)
+        // abi and byteCode are extracted by compiling caver-java-boilerplate/resources/KVstore.sol
         String abi = "[\n" +
                 "  {\n" +
                 "    \"constant\":true,\n" +
@@ -127,26 +126,23 @@ public class Boilerplate {
                 "    \"type\":\"constructor\"\n" +
                 "  }\n" +
                 "]";
-        // You can get contract address
-        // by running caver-java-examples/contract/deploy scenario.
-        String contractAddress = "0x{contractAddress}";
-        SingleKeyring deployerKeyring = caver.wallet.keyring.create(
-                deployerAddress,
-                deployerPrivateKey
-        );
+        String byteCode = "608060405234801561001057600080fd5b506040516106763803806106768339810180604052602081101561003357600080fd5b81019080805164010000000081111561004b57600080fd5b8281019050602081018481111561006157600080fd5b815185600182028301116401000000008211171561007e57600080fd5b5050929190505050806001908051906020019061009c9291906100a3565b5050610148565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f106100e457805160ff1916838001178555610112565b82800160010185558215610112579182015b828111156101115782518255916020019190600101906100f6565b5b50905061011f9190610123565b5090565b61014591905b80821115610141576000816000905550600101610129565b5090565b90565b61051f806101576000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c8063693ec85e1461003b578063e942b5161461016f575b600080fd5b6100f46004803603602081101561005157600080fd5b810190808035906020019064010000000081111561006e57600080fd5b82018360208201111561008057600080fd5b803590602001918460018302840111640100000000831117156100a257600080fd5b91908080601f016020809104026020016040519081016040528093929190818152602001838380828437600081840152601f19601f8201169050808301925050505050505091929192905050506102c1565b6040518080602001828103825283818151815260200191508051906020019080838360005b83811015610134578082015181840152602081019050610119565b50505050905090810190601f1680156101615780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b6102bf6004803603604081101561018557600080fd5b81019080803590602001906401000000008111156101a257600080fd5b8201836020820111156101b457600080fd5b803590602001918460018302840111640100000000831117156101d657600080fd5b91908080601f016020809104026020016040519081016040528093929190818152602001838380828437600081840152601f19601f8201169050808301925050505050505091929192908035906020019064010000000081111561023957600080fd5b82018360208201111561024b57600080fd5b8035906020019184600183028401116401000000008311171561026d57600080fd5b91908080601f016020809104026020016040519081016040528093929190818152602001838380828437600081840152601f19601f8201169050808301925050505050505091929192905050506103cc565b005b60606000826040518082805190602001908083835b602083106102f957805182526020820191506020810190506020830392506102d6565b6001836020036101000a03801982511681845116808217855250505050505090500191505090815260200160405180910390208054600181600116156101000203166002900480601f0160208091040260200160405190810160405280929190818152602001828054600181600116156101000203166002900480156103c05780601f10610395576101008083540402835291602001916103c0565b820191906000526020600020905b8154815290600101906020018083116103a357829003601f168201915b50505050509050919050565b806000836040518082805190602001908083835b6020831061040357805182526020820191506020810190506020830392506103e0565b6001836020036101000a0380198251168184511680821785525050505050509050019150509081526020016040518091039020908051906020019061044992919061044e565b505050565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f1061048f57805160ff19168380011785556104bd565b828001600101855582156104bd579182015b828111156104bc5782518255916020019190600101906104a1565b5b5090506104ca91906104ce565b5090565b6104f091905b808211156104ec5760008160009055506001016104d4565b5090565b9056fea165627a7a72305820cda9307ff998ca73df4fb6024fbb9bff5f5b2fc6293799b12f945d12c6f0d1960029";
+        Contract contract = caver.contract.create(abi);
+        contract.getMethods().forEach((name, method) ->{
+            System.out.println(method.getType() + " " +  caver.abi.buildFunctionString(method));
+        });
+
+        SingleKeyring deployerKeyring = caver.wallet.keyring.create(deployerAddress, deployerPrivateKey);
         caver.wallet.add(deployerKeyring);
 
-        Contract contract = caver.contract.create(abi, contractAddress);
-        SendOptions sendOptions = new SendOptions(deployerKeyring.getAddress(), BigInteger.valueOf(400000));
-        TransactionReceipt.TransactionReceiptData transactionReceiptData = contract.send(
-                sendOptions,
-                "set",
-                "k1",
-                "v1"
-        );
-        System.out.println(objectToString(transactionReceiptData));
+        SendOptions sendOptions = new SendOptions(deployerKeyring.getAddress(), BigInteger.valueOf(650000));
+        Contract deployedContract = contract.deploy(sendOptions, byteCode, "Just", "Test");
+        System.out.println("Deployed address of contract: " + deployedContract.getContractAddress());
 
-        List<Type> callResult = contract.call("get", "k1");
-        System.out.println("Result of calling get function with key:");
-        System.out.println(objectToString(callResult));    }
+        TransactionReceipt.TransactionReceiptData receiptData = deployedContract.send(sendOptions, "set", "Just", "Test");
+        System.out.println(objectToString(receiptData));
+
+        List<Type> returnedData = deployedContract.call("get", "Just");
+        System.out.println(objectToString(returnedData));
+    }
 }
