+++ account/update_account_with_account_key_role_based/build.gradle
@@ -24,7 +24,7 @@ dependencies {
 }
 
 task run(type: JavaExec) {
-    group = "Boilerplate"
+    group = "CaverExample"
     classpath = sourceSets.main.runtimeClasspath
-    mainClass.set("Boilerplate")
+    mainClass.set("CaverExample")
 }
