+++ account/update_account_with_account_key_weighted_multisig/src/main/java/CaverExample.java
@@ -12,7 +12,7 @@ import com.klaytn.caver.transaction.response.PollingTransactionReceiptProcessor;
 import com.klaytn.caver.transaction.response.TransactionReceiptProcessor;
 import com.klaytn.caver.transaction.type.AccountUpdate;
 import com.klaytn.caver.transaction.type.ValueTransfer;
-import com.klaytn.caver.wallet.keyring.RoleBasedKeyring;
+import com.klaytn.caver.wallet.keyring.MultipleKeyring;
 import com.klaytn.caver.wallet.keyring.SingleKeyring;
 import io.github.cdimascio.dotenv.Dotenv;
 import okhttp3.Credentials;
@@ -23,14 +23,13 @@ import java.math.BigInteger;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
-import java.util.List;
 
 /**
- * Boilerplate code about "How to Update Klaytn Account Keys with Caver #3 — AccountKeyRoleBased"
- * Related article - Korean: https://medium.com/klaytn/caver-caver%EB%A1%9C-klaytn-%EA%B3%84%EC%A0%95%EC%9D%98-%ED%82%A4%EB%A5%BC-%EB%B0%94%EA%BE%B8%EB%8A%94-%EB%B0%A9%EB%B2%95-3-accountkeyrolebased-88c20b405f18
- * Related article - English: https://medium.com/klaytn/caver-how-to-update-klaytn-account-keys-with-caver-3-accountkeyrolebased-eb06433ff8da
+ * Example code about "How to Update Klaytn Account Keys with Caver #2 — AccountKeyWeightedMultiSig"
+ * Related article - Korean: https://medium.com/klaytn/caver-caver%EB%A1%9C-klaytn-%EA%B3%84%EC%A0%95%EC%9D%98-%ED%82%A4%EB%A5%BC-%EB%B0%94%EA%BE%B8%EB%8A%94-%EB%B0%A9%EB%B2%95-2-accountkeyweightedmultisig-c317a785299
+ * Related article - English: https://medium.com/klaytn/caver-how-to-update-klaytn-account-keys-with-caver-2-accountkeyweightedmultisig-ed897b4e5b5b
  */
-public class Boilerplate {
+public class CaverExample {
     // You can directly input values for the variables below, or you can enter values in the caver-java-examples/.env file.
     private static String nodeApiUrl = ""; // e.g. "https://node-api.klaytnapi.com/v1/klaytn";
     private static String accessKeyId = ""; // e.g. "KASK1LVNO498YT6KJQFUPY8S";
@@ -38,8 +37,7 @@ public class Boilerplate {
     private static String chainId = ""; // e.g. "1001" or "8217";
     private static String senderAddress = ""; // e.g. "0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199"
     private static String senderPrivateKey = ""; // e.g. "0x42f6375b608c2572fadb2ed9fd78c5c456ca3aa860c43192ad910c3269727fc7"
-    private static String recipientAddress = ""; // e.g. "0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199"
-
+    private static String recipientAddress= ""; // e.g. "0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199"
 
     public static void main(String[] args) {
         try {
@@ -78,7 +76,7 @@ public class Boilerplate {
     }
 
     public static void run() throws Exception {
-        System.out.println("=====> Update AccountKey to AccountKeyRoleBased");
+        System.out.println("=====> Update AccountKey to AccountKeyWeightedMultiSig");
 
         HttpService httpService = new HttpService(nodeApiUrl);
         httpService.addHeader("Authorization", Credentials.basic(accessKeyId, secretAccessKey));
@@ -89,30 +87,24 @@ public class Boilerplate {
         SingleKeyring senderKeyring = caver.wallet.keyring.create(senderAddress, senderPrivateKey);
         caver.wallet.add(senderKeyring);
 
-        List<String[]> newRoleBasedKeys = caver.wallet.keyring.generateRolBasedKeys(new int[]{2, 1, 3});
-        System.out.println("new private keys by role: " + objectToString(newRoleBasedKeys));
+        // Create new private keys
+        String[] newKeys = caver.wallet.keyring.generateMultipleKeys(3);
+        System.out.println("new private keys: " + objectToString(newKeys));
 
-        // Create new Keyring instance as RoleBasedKeyring with new private keys by role
-        RoleBasedKeyring newKeyring = caver.wallet.keyring.create(senderKeyring.getAddress(), newRoleBasedKeys);
-        BigInteger[][] optionWeight = {
-                {BigInteger.ONE, BigInteger.ONE},
-                {},
-                {BigInteger.valueOf(2), BigInteger.ONE, BigInteger.ONE}
-        };
-        WeightedMultiSigOptions[] options = {
-                new WeightedMultiSigOptions(BigInteger.valueOf(2), Arrays.asList(optionWeight[0])),
-                new WeightedMultiSigOptions(),
-                new WeightedMultiSigOptions(BigInteger.valueOf(3), Arrays.asList(optionWeight[2]))
-        };
-        // Create an Account instance that includes the address and the role based key
-        Account account = newKeyring.toAccount(Arrays.asList(options));
+        // Create new Keyring as MultipleKeyring instance with new private keys
+        MultipleKeyring newKeyring = caver.wallet.keyring.create(senderKeyring.getAddress(), newKeys);
+        BigInteger[] weights = {BigInteger.valueOf(2), BigInteger.ONE, BigInteger.ONE};
+        WeightedMultiSigOptions options = new WeightedMultiSigOptions(BigInteger.valueOf(3), Arrays.asList(weights));
+        // Create an Account instance that includes the address and the weighted multisig key
+        Account account = newKeyring.toAccount(options);
+        System.out.println(objectToString(account));
 
         // Create account update transaction instance
         AccountUpdate accountUpdate = caver.transaction.accountUpdate.create(
                 TxPropertyBuilder.accountUpdate()
                         .setFrom(senderKeyring.getAddress())
                         .setAccount(account)
-                        .setGas(BigInteger.valueOf(150000))
+                        .setGas(BigInteger.valueOf(100000))
         );
 
         // Sign the transaction
@@ -130,7 +122,7 @@ public class Boilerplate {
 
         // Get accountKey from network
         AccountKey accountKey = caver.rpc.klay.getAccountKey(senderKeyring.getAddress()).send();
-        System.out.println("Result of account key update to AccountKeyRoleBased");
+        System.out.println("Result of account key update to AccountKeyWeightedMultiSig");
         System.out.println("Account address: " + senderKeyring.getAddress());
         System.out.println("accountKey => ");
         System.out.println(objectToString(accountKey));
@@ -143,16 +135,16 @@ public class Boilerplate {
                         .setFrom(senderKeyring.getAddress())
                         .setTo(recipientAddress)
                         .setValue(BigInteger.valueOf(1))
-                        .setGas(BigInteger.valueOf(150000))
+                        .setGas(BigInteger.valueOf(100000))
         );
 
         // Sign the transaction with updated keyring
-        // This sign function will sign the transaction with all private keys in RoleTrasnsactionKey in the keyring
+        // This sign function will sign the transaction with all private keys in the keyring
         caver.wallet.sign(senderKeyring.getAddress(), vt);
         // Send transaction
         Bytes32 vtResult = caver.rpc.klay.sendRawTransaction(vt).send();
         TransactionReceipt.TransactionReceiptData vtReceiptData = receiptProcessor.waitForTransactionReceipt(vtResult.getResult());
-        System.out.println("After account update value transfer transaction receipt => ");
+        System.out.println("Receipt of value transfer transaction after account update => ");
         System.out.println(objectToString(vtReceiptData));
     }
 }
