+++ setup.sh
@@ -9,7 +9,7 @@ PROJECT_ROOT_DIR=$PWD
 ASSETS_DIR=$PROJECT_ROOT_DIR/assets
 BUILD_GRADLE="build.gradle"
 BOILER_PLATE_BUILD_GRADLE=$ASSETS_DIR/buildGradle
-BOILER_PLATE_TEMPLATE=$ASSETS_DIR/Boilerplate.java
+BOILER_PLATE_TEMPLATE=$ASSETS_DIR/CaverExample.java
 NEW_SCENARIO=$1
 NEW_SCENARIO_SRC_ROOT=$NEW_SCENARIO/src/main/java
