+++ transaction/send_klay_with_keystore_file/src/main/java/CaverExample.java
@@ -2,32 +2,36 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.klaytn.caver.Caver;
+import com.klaytn.caver.methods.response.TransactionReceipt;
 import com.klaytn.caver.transaction.TxPropertyBuilder;
+import com.klaytn.caver.transaction.response.PollingTransactionReceiptProcessor;
 import com.klaytn.caver.transaction.type.ValueTransfer;
 import com.klaytn.caver.utils.Utils;
-import com.klaytn.caver.wallet.keyring.SingleKeyring;
+import com.klaytn.caver.wallet.keyring.AbstractKeyring;
+import com.klaytn.caver.wallet.keyring.KeyStore;
 import io.github.cdimascio.dotenv.Dotenv;
 import okhttp3.Credentials;
+import org.web3j.protocol.ObjectMapperFactory;
 import org.web3j.protocol.http.HttpService;
 
+import java.io.File;
+import java.io.FileNotFoundException;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
 /**
- * Boilerplate code about "How to use Klay Units."
- * Related reference - Korean: https://ko.docs.klaytn.com/klaytn/design/klaytn-native-coin-klay#units-of-klay
- * Related reference - English: https://docs.klaytn.com/klaytn/design/klaytn-native-coin-klay#units-of-klay
+ * Example code about "How to send KLAY with keystore file."
+ * Related article - Korean: https://medium.com/klaytn/common-architecture-of-caver-f7a7a1c554de
+ * Related article - English: https://medium.com/klaytn/common-architecture-of-caver-a714224a0047
  */
-public class Boilerplate {
+public class CaverExample {
     // You can directly input values for the variables below, or you can enter values in the caver-java-examples/.env file.
     private static String nodeApiUrl = ""; // e.g. "https://node-api.klaytnapi.com/v1/klaytn";
     private static String accessKeyId = ""; // e.g. "KASK1LVNO498YT6KJQFUPY8S";
     private static String secretAccessKey = ""; // e.g. "aP/reVYHXqjw3EtQrMuJP4A3/hOb69TjnBT3ePKG";
     private static String chainId = ""; // e.g. "1001" or "8217";
-    private static String senderAddress = ""; // e.g. "0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199"
-    private static String senderPrivateKey = ""; // e.g. "0x42f6375b608c2572fadb2ed9fd78c5c456ca3aa860c43192ad910c3269727fc7"
     private static String recipientAddress = ""; // e.g. "0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199"
 
     public static void main(String[] args) {
@@ -61,58 +65,50 @@ public class Boilerplate {
         accessKeyId = accessKeyId.equals("") ? env.get("ACCESS_KEY_ID") : accessKeyId;
         secretAccessKey = secretAccessKey.equals("") ? env.get("SECRET_ACCESS_KEY") : secretAccessKey;
         chainId = chainId.equals("") ? env.get("CHAIN_ID") : chainId;
-        senderAddress = senderAddress.equals("") ? env.get("SENDER_ADDRESS") : senderAddress;
-        senderPrivateKey = senderPrivateKey.equals("") ? env.get("SENDER_PRIVATE_KEY") : senderPrivateKey;
         recipientAddress = recipientAddress.equals("") ? env.get("RECIPIENT_ADDRESS") : recipientAddress;
     }
 
-    public static void run() {
+    public static void run() throws Exception {
         HttpService httpService = new HttpService(nodeApiUrl);
         httpService.addHeader("Authorization", Credentials.basic(accessKeyId, secretAccessKey));
         httpService.addHeader("x-chain-id", chainId);
         Caver caver = new Caver(httpService);
 
-        SingleKeyring senderKeyring = caver.wallet.keyring.create(senderAddress, senderPrivateKey);
+        // 1. Create your own keystore file at "https://baobab.wallet.klaytn.com/create".
+        // 2. Rename that keystore file name with `keystore.json` or change the filename in line 81 and 84.
+        // 3. Place that keystore file at `caver-java-examples/kct/deploy_kip7_token_contract_with_keystore_file/resources`.
+        // 4. Get 5 KLAY at "https://baobab.wallet.klaytn.com/faucet".
+        File file = new File("resources/keystore.json");
+        if(file.exists() == false) {
+            // Handles when you run this CaverExample as sub-module using IDE.
+            file = new File("transaction/send_klay_with_keystore_file/resources/keystore.json");
+            if(file.exists() == false) {
+                throw new FileNotFoundException("Cannot find keystore.json file.");
+            }
+        }
+        String password = ""; // Put your password here.
+        ObjectMapper objectMapper = ObjectMapperFactory.getObjectMapper();
+        KeyStore keyStore = objectMapper.readValue(file, KeyStore.class);
+        AbstractKeyring senderKeyring = caver.wallet.keyring.decrypt(keyStore, password);
         caver.wallet.add(senderKeyring);
 
-        // Because a field "value" always interprets its value as a unit "peb",
-        // you must take care what is the actual value when you sending some KLAY.
+        // Send 1 KLAY.
         ValueTransfer vt = caver.transaction.valueTransfer.create(
                 TxPropertyBuilder.valueTransfer()
                         .setFrom(senderKeyring.getAddress())
                         .setTo(recipientAddress)
+                        .setValue(new BigInteger(
+                                caver.utils.convertToPeb(
+                                        BigDecimal.valueOf(1),
+                                        Utils.KlayUnit.KLAY
+                                )
+                        ))
                         .setGas(BigInteger.valueOf(25000))
-                        .setValue(BigInteger.ONE)
         );
-        // Example-1: Sending 0.5 KLAY to recipient
-        // option-1 (Recommended): Use KlayUnit.
-        vt.setValue(new BigDecimal(caver.utils.convertToPeb("0.5", Utils.KlayUnit.KLAY)).toBigInteger());
-        System.out.println("Example-1) The value what we set using option-1 is " + vt.getValue());
-
-        // 1 KLAY is actually 10^18(=1000000000000000000) peb. So if you want send 0.5 KLAY,
-        // option-2 (Not recommended): Set actual peb value directly to ValueTransfer transaction instance.
-        vt.setValue(BigInteger.valueOf(500000000000000000L)); // 5 * (10^17)
-        System.out.println("Example-1) The value what we set using option-2 is " + vt.getValue());
-
-        // Example-2: Sending 0.05 KLAY to recipient
-        // option-1 (Recommended): Use KlayUnit.
-        vt.setValue(new BigDecimal(caver.utils.convertToPeb("0.05", Utils.KlayUnit.KLAY)).toBigInteger());
-        System.out.println("Example-2) The value what we set using option-1 is " + vt.getValue());
-
-        // 1 KLAY is actually 10^18(=1000000000000000000) peb. So if you want send 0.05 KLAY,
-        // option-2 (Not recommended): Set actual peb value directly to ValueTransfer transaction instance.
-        vt.setValue(BigInteger.valueOf(50000000000000000L)); // 5 * (10^16)
-        System.out.println("Example-2) The value what we set using option-2 is " + vt.getValue());
-
-        // Example-3: Sending 0.005 KLAY to recipient
-        // option-1 (Recommended): Use KlayUnit.
-        vt.setValue(new BigInteger(caver.utils.convertToPeb(BigDecimal.valueOf(5), Utils.KlayUnit.mKLAY)));
-        System.out.println("Example-3) The value what we set using option-1 is " + vt.getValue());
-
-        // 1 KLAY is actually 10^18(=1000000000000000000) peb. So if you want send 0.005 KLAY,
-        // option-2 (Not recommended): Set actual peb value directly to ValueTransfer transaction instance.
-        vt.setValue(BigInteger.valueOf(5000000000000000L)); // 5 * (10^15)
-        System.out.println("Example-3) The value what we set using option-2 is " + vt.getValue());
-
+        caver.wallet.sign(senderKeyring.getAddress(), vt);
+        String txHash = caver.rpc.klay.sendRawTransaction(vt).send().getResult();
+        PollingTransactionReceiptProcessor receiptProcessor = new PollingTransactionReceiptProcessor(caver, 1000, 15);
+        TransactionReceipt.TransactionReceiptData receiptData = receiptProcessor.waitForTransactionReceipt(txHash);
+        System.out.println(objectToString(receiptData));
     }
 }
