+++ utils/play_with_units_in_klaytn/src/main/java/CaverExample.java
@@ -2,10 +2,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.klaytn.caver.Caver;
-import com.klaytn.caver.methods.response.Bytes32;
-import com.klaytn.caver.methods.response.TransactionReceipt;
 import com.klaytn.caver.transaction.TxPropertyBuilder;
-import com.klaytn.caver.transaction.response.PollingTransactionReceiptProcessor;
 import com.klaytn.caver.transaction.type.ValueTransfer;
 import com.klaytn.caver.utils.Utils;
 import com.klaytn.caver.wallet.keyring.SingleKeyring;
@@ -19,11 +16,11 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 
 /**
- * BoilerPlate code about "How to send Klay."
- * Related reference - Korean: https://ko.docs.klaytn.com/bapp/sdk/caver-java/getting-started#sending-klay-at-a-glance
- * Related reference - English: https://docs.klaytn.com/bapp/sdk/caver-java/getting-started#sending-klay-at-a-glance
+ * Example code about "How to use Klay Units."
+ * Related reference - Korean: https://ko.docs.klaytn.com/klaytn/design/klaytn-native-coin-klay#units-of-klay
+ * Related reference - English: https://docs.klaytn.com/klaytn/design/klaytn-native-coin-klay#units-of-klay
  */
-public class Boilerplate {
+public class CaverExample {
     // You can directly input values for the variables below, or you can enter values in the caver-java-examples/.env file.
     private static String nodeApiUrl = ""; // e.g. "https://node-api.klaytnapi.com/v1/klaytn";
     private static String accessKeyId = ""; // e.g. "KASK1LVNO498YT6KJQFUPY8S";
@@ -69,7 +66,7 @@ public class Boilerplate {
         recipientAddress = recipientAddress.equals("") ? env.get("RECIPIENT_ADDRESS") : recipientAddress;
     }
 
-    public static void run() throws Exception {
+    public static void run() {
         HttpService httpService = new HttpService(nodeApiUrl);
         httpService.addHeader("Authorization", Credentials.basic(accessKeyId, secretAccessKey));
         httpService.addHeader("x-chain-id", chainId);
@@ -78,23 +75,44 @@ public class Boilerplate {
         SingleKeyring senderKeyring = caver.wallet.keyring.create(senderAddress, senderPrivateKey);
         caver.wallet.add(senderKeyring);
 
-        // Send 1 KLAY.
+        // Because a field "value" always interprets its value as a unit "peb",
+        // you must take care what is the actual value when you sending some KLAY.
         ValueTransfer vt = caver.transaction.valueTransfer.create(
                 TxPropertyBuilder.valueTransfer()
-                        .setFrom(senderAddress)
+                        .setFrom(senderKeyring.getAddress())
                         .setTo(recipientAddress)
-                        .setValue(new BigInteger(
-                                caver.utils.convertToPeb(
-                                        BigDecimal.valueOf(1),
-                                        Utils.KlayUnit.KLAY
-                                )
-                        ))
                         .setGas(BigInteger.valueOf(25000))
+                        .setValue(BigInteger.ONE)
         );
-        caver.wallet.sign(senderKeyring.getAddress(), vt);
-        String txHash = caver.rpc.klay.sendRawTransaction(vt).send().getResult();
-        PollingTransactionReceiptProcessor receiptProcessor = new PollingTransactionReceiptProcessor(caver, 1000, 15);
-        TransactionReceipt.TransactionReceiptData receiptData = receiptProcessor.waitForTransactionReceipt(txHash);
-        System.out.println(objectToString(receiptData));
+        // Example-1: Sending 0.5 KLAY to recipient
+        // option-1 (Recommended): Use KlayUnit.
+        vt.setValue(new BigDecimal(caver.utils.convertToPeb("0.5", Utils.KlayUnit.KLAY)).toBigInteger());
+        System.out.println("Example-1) The value what we set using option-1 is " + vt.getValue());
+
+        // 1 KLAY is actually 10^18(=1000000000000000000) peb. So if you want send 0.5 KLAY,
+        // option-2 (Not recommended): Set actual peb value directly to ValueTransfer transaction instance.
+        vt.setValue(BigInteger.valueOf(500000000000000000L)); // 5 * (10^17)
+        System.out.println("Example-1) The value what we set using option-2 is " + vt.getValue());
+
+        // Example-2: Sending 0.05 KLAY to recipient
+        // option-1 (Recommended): Use KlayUnit.
+        vt.setValue(new BigDecimal(caver.utils.convertToPeb("0.05", Utils.KlayUnit.KLAY)).toBigInteger());
+        System.out.println("Example-2) The value what we set using option-1 is " + vt.getValue());
+
+        // 1 KLAY is actually 10^18(=1000000000000000000) peb. So if you want send 0.05 KLAY,
+        // option-2 (Not recommended): Set actual peb value directly to ValueTransfer transaction instance.
+        vt.setValue(BigInteger.valueOf(50000000000000000L)); // 5 * (10^16)
+        System.out.println("Example-2) The value what we set using option-2 is " + vt.getValue());
+
+        // Example-3: Sending 0.005 KLAY to recipient
+        // option-1 (Recommended): Use KlayUnit.
+        vt.setValue(new BigInteger(caver.utils.convertToPeb(BigDecimal.valueOf(5), Utils.KlayUnit.mKLAY)));
+        System.out.println("Example-3) The value what we set using option-1 is " + vt.getValue());
+
+        // 1 KLAY is actually 10^18(=1000000000000000000) peb. So if you want send 0.005 KLAY,
+        // option-2 (Not recommended): Set actual peb value directly to ValueTransfer transaction instance.
+        vt.setValue(BigInteger.valueOf(5000000000000000L)); // 5 * (10^15)
+        System.out.println("Example-3) The value what we set using option-2 is " + vt.getValue());
+
     }
 }
